+++ src/core/moves/admixture/AdmixtureEdgeDivergenceMerge.h
@@ -34,7 +34,7 @@ namespace RevBayesCore {
         // Basic utility functions
         AdmixtureEdgeDivergenceMerge*   clone(void) const;                                                                  //!< Clone object
         void                            swapNode(DagNode *oldN, DagNode *newN);
-        bool                            isActive(int g) const;
+        bool                            isActive(unsigned long g) const;
         
     protected:
         const std::string&              getMoveName(void) const;                                                            //!< Get the name of the move for summary printing
